fix(edge-doctor): insulate doctor tests from inherited CORDUM_AGENTD_FAIL_CLOSED
--fail-closed defaults from CORDUM_AGENTD_FAIL_CLOSED, so an inherited
"false" made the policy-mode check warn and flipped the doctor exit
code to 2. TestEdgeDoctorSuccessJSON then failed on any machine that
exported the variable -- likely exactly the people working on edge.

Clear it in runEdgeDoctorForTest alongside the existing CORDUM_API_KEY
hygiene, and invoke runEdgeDoctorCmd directly in the one test that
exercises the env default.

Also split the posture helpers into edge_doctor_posture.go: the additions
pushed edge_doctor_checks.go to 340 lines, past the 300-line limit.

task-5e027c46
